+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## 1.0.12
+## 1.1
 
 ### New features
 - `GetEntityAfterCreate` method added in IRestCollection. It defines how the created entity should be returned to the client. Useful when the entity is not persisted in DB.